refactor(core): extract ImagePickerService to lib/core/services (FR-EX-05)
Per architect notes section 2.4, moves the existing image picker
abstraction from lib/features/auth/data/image_picker_service.dart to
lib/core/services/image_picker_service.dart so the expense feature
can import it for the new receipt picker flow without an
auth->expenses dependency direction.

File contents unchanged (git mv preserves blame). Updates the two
production callers (profile_setup_controller, edit_profile_controller)
and the five test files referencing the old path. Adds a new smoke
test at test/core/services/image_picker_service_test.dart that
confirms the imagePickerServiceProvider resolves to a
DefaultImagePickerService by default and can be overridden with a
fake for tests.

No behaviour change. 71 affected tests still pass.
