+++ test/features/profile/edit_profile_screen_test.dart
@@ -5,7 +5,7 @@ import 'package:image_picker/image_picker.dart';
 
 import 'package:onebytwo/features/auth/application/analytics_provider.dart';
 import 'package:onebytwo/features/auth/application/auth_state_provider.dart';
-import 'package:onebytwo/features/auth/data/image_picker_service.dart';
+import 'package:onebytwo/core/services/image_picker_service.dart';
 import 'package:onebytwo/features/auth/data/user_repository.dart';
 import 'package:onebytwo/features/auth/domain/auth_state.dart';
 import 'package:onebytwo/features/auth/domain/user_model.dart';
